# api/api.yml
openapi: 3.0.3
info:
  title: Go API POC
  description: A proof of concept API demonstrating some Go tools and AWS hosting options.
  version: 0.9.0
servers:
  - url: https://devcatalog.giftbit.com
paths:
...
  /go/reward:
    get:
      summary: Get a list of active rewards for your account
      responses:
        '200':
          description: A list of rewards
          content:
            application/json:
              schema:
                type: object
                properties:
                  rewards:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reward'
    post:
      summary: Create a reward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RewardCreation'
      responses:
        '201':
          description: The created reward
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reward'
#   /go/item/{id}:
#     get:
#       summary: Get an item by ID
#       parameters:
#         - name: id
#           in: path
#           required: true
#           description: The ID of the item
#           schema:
#             type: string
#       responses:
#         '200':
#           description: An item
#           content:
#             application/json:
#               schema:
#                 $ref: '#/components/schemas/Item'
#     put:
#       summary: Update an item by ID
#       parameters:
#         - name: id
#           in: path
#           required: true
#           description: The ID of the item
#           schema:
#             type: string
#       requestBody:
#         required: true
#         content:
#           application/json:
#             schema:
#               $ref: '#/components/schemas/Item'
#       responses:
#         '200':
#           description: The updated item
#           content:
#             application/json:
#               schema:
#                 $ref: '#/components/schemas/Item'
#     delete:
#       summary: Delete an item by ID
#       parameters:
#         - name: id
#           in: path
#           required: true
#           description: The ID of the item
#           schema:
#             type: string
#       responses:
#         '204':
#           description: The item was deleted
components:
  schemas:
    Reward:
      type: object
      properties:
        id:
          type: number
          example: 1
        brand:
          type: string
          example: Pokemart
        currency:
          type: string
          example: PKD
        denomination:
          type: number
          example: 1000
    RewardCreation:
      type: object
      properties:
        brand:
          type: string
          example: Pokemart
        currency:
          type: string
          example: PKD
        denomination:
          type: number
          example: 1000